+++ integration/app/list/list.module.ts
@@ -6,9 +6,11 @@ import { NgxsModule } from '@ngxs/store';
 import { ListState } from '@integration/list/list.state';
 import { ListComponent } from '@integration/list/list.component';
 import { routes } from '@integration/list/list.routes';
+import { ListResolver } from '@integration/list/list.resolver';
 
 @NgModule({
   imports: [CommonModule, RouterModule.forChild(routes), NgxsModule.forFeature([ListState])],
-  declarations: [ListComponent]
+  declarations: [ListComponent],
+  providers: [ListResolver]
 })
 export class ListModule {}