+++ packages/store/tests/default-states.spec.ts
@@ -1,9 +1,10 @@
-import { Action, NgxsModule, State, StateContext, Store } from '../src/public_api';
 import { TestBed } from '@angular/core/testing';
+import { APP_INITIALIZER, NgModule } from '@angular/core';
+
+import { Action, NgxsModule, State, StateContext, Store } from '../src/public_api';
+import { StateStream } from '../src/internal/state-stream';
 
 describe('Reusable States', () => {
-  let store: Store;
-
   class UpdateFoo {
     static readonly type = '[update] foo';
     constructor(public payload: number) {}
@@ -31,7 +32,7 @@ describe('Reusable States', () => {
       ]
     });
 
-    store = TestBed.get(Store);
+    const store = TestBed.get(Store);
 
     let stateValue = store.selectSnapshot(FooState);
     expect(stateValue).toEqual([4, 5, 6]);
@@ -39,6 +40,43 @@ describe('Reusable States', () => {
     store.dispatch(new UpdateFoo(7));
     stateValue = store.selectSnapshot(FooState);
     expect(stateValue).toEqual([4, 5, 6, 7]);
+  });
+
+  it('should not overwrite a state stream that is already initialised', () => {
+    @NgModule({
+      providers: [
+        {
+          provide: APP_INITIALIZER,
+          useFactory: (stateStream: StateStream) => {
+            stateStream.next({ foo: [1, 2, 3] });
+            return () => {};
+          },
+          multi: true,
+          deps: [StateStream]
+        }
+      ]
+    })
+    class MyStateSettingModule {}
+
+    TestBed.configureTestingModule({
+      imports: [
+        MyStateSettingModule,
+        NgxsModule.forRoot([FooState], {
+          defaultsState: {
+            foo: [4, 5, 6]
+          }
+        })
+      ]
+    });
+
+    const store = TestBed.get(Store);
+
+    let stateValue = store.selectSnapshot(FooState);
+    expect(stateValue).toEqual([1, 2, 3]);
+
+    store.dispatch(new UpdateFoo(7));
+    stateValue = store.selectSnapshot(FooState);
+    expect(stateValue).toEqual([1, 2, 3, 7]);
   });
 
   it('should be correct readonly state', () => {
@@ -56,7 +94,7 @@ describe('Reusable States', () => {
       ]
     });
 
-    store = TestBed.get(Store);
+    const store = TestBed.get(Store);
 
     expect(store.snapshot()).toEqual({
       configState: { a: 1, b: 2 },